correções e ajuste de código

+++ src/main/resources/db/migration/V02__criar_e_registrar_livro.sql
@@ -3,7 +3,7 @@ id BIGINT PRIMARY KEY AUTO_INCREMENT,
 nome_autor VARCHAR(50) NOT NULL,
 titulo VARCHAR(255) NOT NULL,
 texto VARCHAR(1000) NOT NULL,
-categoria int(10) NOT NULL
+categoria BIGINT(10) NOT NULL
 )engine = InnoDB charset=utf8;
 
 INSERT INTO `booksbuild`.livro (nome_autor,titulo,texto,categoria) values ('Robert Martin','Clean Code','Lorem ipsum',1);
